refactor tweets
since there will be two different streams of tweets, one for the homepage, and one fore the news page, there are now two different resouce end points for each one of those. Also added a new worker to reflect this, which is much like the previous twitter worker. Although the  functionality is the same as the  functionality. That will happen next