tumbls & tweets realiably
Updated the way the Tumbls and Tweets are gathered. All tumbls with the particular account will be gathered and diffed. 20 at a time, which is the API max. For now, there will be no delay between calls. But may need to down the road. Tweets are called for, and diffed. Since we can't realiably get all tweets, we will only store the ones that we can get. Which means we don't have any  permalinks to tweets on our news page.