geo form flow and backend
set up the dropdownConditionalText to be used in formFlow as the select_geo component. This now grabs data in the way that the backend will anticipate, in order to populate a TopLevelGeo object. The data on the frontend is stored in the user object, which can be used in the profile, and parities the data coming from the server. User data is round tripable.