radio element interface
some functions pulled strings, others pulled objects with the string values in it. noramlized things, and now  and  both accept strings and objects as arguments. either way, the point of the interface is to define a selection WITHIN the data provided. sets of data being passed in are small enough that looping is no problem.

also reworked structure of code, and removed code that is no longer being used