modal animation
we have a functioning system, that is delightful in a couple of ways. the filtering mechanism, and the network mechanism. but the sing up form is a little blah. it needs to be a bit more responsive. and could use a bit of feedback as to where you are in the process, and to top it off, it could use a constant object. something that moves with you as you are making your way through the form.

introducing a network graph to get at that. we'll see what it looks like in the end