form flow fixes
top_level_input was not nullable, and set_geo was looking for it to be. similar problem with trying to update work_in_ values for a steamie model with no work_in value.

PUT http requests were choking on tastypie, and realized that the PATCH requests could be used, since we have a user id to submit against at that point.

changed interface into user data on the frontend to be normalized. no guessing if there it is stuffed away in data.objects[0]. Keeping that structure was previously used for the PUT requests that were choking.