responsiveness
refactored the profiles. since the only two things that need to be validated are consistent on both, the validation scheme is the same. its more of a validation + has anything been updated function. if so, enable save. pulled all of that out into the profile module.
added cancel button to the sign up form