prefixed css
Browsers that support transitions and animations can now make it around the site.

Ended up not turning on the animations for , ,  and  for performance reasons. Only  handles it well. The only animations that are being used at the moment are to rotate the SVG args, and to pulse the SVG loading orb.