Fixes #35
Updated the change_work_in_count method of the TLG model to not have a default field. It was previously set to have a default feild of education.
Updated the reset_tlg_count management command to log which TLGs required a revised work_in_total based on having counted all steamies on the TLG. In my one run of it so far, the only discrepencies were on work_in_education totals. I mentioned above that the function in charge of changing totals, was previously defaulting to changing education, with a value of 1. This leads me to believe that the default value of this function was leading to the inaccurate TLG counts. reset_tlg_count will ensure that is the case. Setting this up to run every 6 hours.
There is also a saved search in Loggly, looking out for instances where a TLG value had to get reset. This will get run daily. If there are occurences of the inaccurate TLG count, this will pick up on it, with a timestamp of when it happened within the 6 hour window between runs. We can then further inspect the logs to get a better another insight on where this is happening.
My hunch is that its fixed, though.